requirements.txt
pandas==2.2.2
pyarrow==16.1.0
SQLAlchemy==2.0.32
langdetect==1.0.9
pycld3==0.22
fasttext==0.9.2
pytest==7.4.3
regex==2023.10.3
ftfy==6.1.1
chardet>=5.2.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
scikit-learn>=1.3.0
numpy>=1.24.0
tqdm>=4.65.0
flask>=2.3.0
flask-cors>=4.0.0
